Added user-error handling system for providers.
Custom errors are designed and exported by each Provider.
Each provider can define its own errorHandler function,
which allows the user to handle the custom errors that appeared
while processing the request.

The provider can do anything it wants with the errorHandler (which
could facilitate its end-user use) as long as it sticks to the
base definitions of ErrorHandler found in the file that contains
the Provider base class.